requirements.txt
anyio==3.7.0
arrow==1.2.3
asks==3.0.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.4.0
comtypes==1.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.12.2
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-secret-manager==2.16.1
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
httplib2==0.22.0
idna==3.4
lxml==4.9.2
mega.py==1.0.8
mysqlclient==2.1.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
pandas==1.5.3
pathlib==1.0.1
proto-plus==1.22.2
protobuf==4.23.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyDrive==1.3.1
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.4.3
selenium-requests==2.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.46
tenacity==5.1.5
tldextract==3.4.4
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wsproto==1.2.0
